Fixes from code review.
