+++ server/fleet/aggregated_stats.go
@@ -1,0 +1,7 @@
+package fleet
+
+const (
+	AggregatedStatsTypeScheduledQuery = "scheduled_query"
+)
+
+type AggregatedStatsType string
